+++ CHANGELOG.md
@@ -5,6 +5,11 @@
     - added ctrl-% shortcut
     - starting to think about Safe Mode
     - simplifying logic to generate and scale product images
+    - fixed BLE Firmware Rev 
+    - added Power Connection State (pending FW fix)
+    - added Laser Max Temp DegC
+    - added AssemblyRevision (packed and expanded)
+    - deprecated Baud Rate and Linearity Coeffs
 - 2026-04-16 4.2.1
     - Bluetooth
         - added RnD.BluetoothTester plugin